pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "Misha4231", email = "mishagalko1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django >=5.2.3,<6.0.0",
    "djangorestframework >=3.16.0,<4.0.0",
    "weasyprint (>=65.1,<66.0)",
    "django-ipware (>=7.0.1,<8.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
